Added some functions in models.py to help clear old games
